refactor: apiUtils, authCalls, endpoints
